+++ projects/admin/src/app/initiatives/initiative-detail/initiative-detail-committee/initiative-detail-committee-members-table/initiative-detail-committee-members-table.component.ts
@@ -74,6 +74,9 @@ export class InitiativeDetailCommitteeMembersTableComponent implements OnChanges
   @Output()
   public resetClicked = new EventEmitter<InitiativeCommitteeMember>();
 
+  @Output()
+  public editClicked = new EventEmitter<InitiativeCommitteeMember>();
+
   protected dataSource = new TableDataSource<InitiativeCommitteeMember>();
 
   public ngOnChanges(): void {