+++ projects/citizen/src/app/app.component.ts
@@ -72,7 +72,7 @@ export class AppComponent implements OnInit, OnDestroy {
   constructor() {
     const stylingService = inject(StylingService);
 
-    stylingService.setTheme(ColorTokensThemes.SchalterELight);
+    stylingService.setTheme(ColorTokensThemes.SGSchalterELight);
     stylingService.setRadius(CornerRadiusTokensThemes.Large);
   }